추가: makgeolli 앱추가, 템플릿추가
