+++ makgeolli/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from makgeolli.models import Makgeolli
+
+# Register your models here.
+admin.site.register(Makgeolli)
