+++ post/admin.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
-from post.models import Comment, Makgeolli
+from post.models import Comment
 
 
 admin.site.register(Comment)
-admin.site.register(Makgeolli)
