+++ post/urls.py
@@ -1,7 +1,5 @@
 from django.urls import path
 from post import views
-from django.conf import settings
-from django.conf.urls.static import static
 
 app_name = 'post'
 
@@ -12,4 +10,4 @@ urlpatterns = [
     path('comment/', views.create_comment, name= 'create_comment'),
     #path('<int:pk>/edit/', views.edit, name = 'edit'),
 
-] + static(settings.MEDIA_URL, document_root = settings.MEDIA_ROOT)
+]
